+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/CasesController.java
@@ -24,7 +24,7 @@ import uk.gov.hmcts.reform.civil.model.citizenui.dto.EventDto;
 import uk.gov.hmcts.reform.civil.model.search.Query;
 import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 import uk.gov.hmcts.reform.civil.service.citizen.events.CaseEventService;
-import uk.gov.hmcts.reform.civil.service.citizen.events.EventSubmissionParams;
+import uk.gov.hmcts.reform.civil.model.EventSubmissionParams;
 
 import static java.util.Collections.emptyList;
