+++ src/main/java/uk/gov/hmcts/reform/civil/model/EventSubmissionParams.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.civil.service.citizen.events;
+package uk.gov.hmcts.reform.civil.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 
 import java.util.Map;
@@ -10,6 +11,7 @@ import java.util.Map;
 @Data
 @Builder
 @AllArgsConstructor
+@NoArgsConstructor
 public class EventSubmissionParams {
 
     private static final String DRAFT_CLAIM_ID = "draft";
