+++ src/main/java/uk/gov/hmcts/reform/civil/service/citizen/events/CaseEventService.java
@@ -9,6 +9,7 @@ import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
+import uk.gov.hmcts.reform.civil.model.EventSubmissionParams;
 import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 
 import static uk.gov.hmcts.reform.civil.CaseDefinitionConstants.GENERAL_APPLICATION_CASE_TYPE;
